+++ components/ListComponents/index.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Tabs, Button, Typography } from 'antd';
 import { useRouter } from 'next/router';
-import { MenuEmptyMessage } from 'common-util/MenuCommon';
+import { ListEmptyMessage } from 'common-util/ListCommon';
 import { EmptyMessage } from '../styles';
 
 const { TabPane } = Tabs;
@@ -36,11 +36,11 @@ const MenuComponents = ({ account, balance }) => {
         )}
       >
         <TabPane tab="All" key="all">
-          <MenuEmptyMessage type="component" />
+          <ListEmptyMessage type="component" />
         </TabPane>
         <TabPane tab="My Components" key="my_components">
           {account ? (
-            <MenuEmptyMessage type="component" />
+            <ListEmptyMessage type="component" />
           ) : (
             <EmptyMessage width="200px">
               To see your components, connect a wallet.
