+++ pages/components/register.jsx
@@ -1,3 +1,3 @@
-import RegisterComponents from 'components/MenuComponents/register';
+import RegisterComponents from 'components/ListComponents/register';
 
 export default RegisterComponents;
